Ensure that lockscreen power menu options the visibility is hidden
This change will make sure that when the switch to change the visibility of the
power menu is toggled off, the options to toggle reboot, screenshot, etc are also
disabled. Leaving them is on is poor UX.

With this change everything will work in harmony. The main switches will override the
lockscreen switches while enabled but the lockscreen visibility switch will have overall
control as it should.